commit after finishing the report

+++ Client/controller.py
@@ -18,19 +18,19 @@ class ControllerClient:
         self.view_client.show_clients(clients)
 
     def update_client(self):
-        # Запрос ID бронирования, которое нужно обновить
+        # Request the ID of the reservation to be updated
         client_id = self.view_client.get_client_id()
 
-        # Проверка, существует ли бронирование с указанным ID
+        # Check if there is a reservation with the specified ID
         client_exists = self.model_client.check_client_existence(client_id)
 
         if client_exists:
-            # Запрос обновленных данных о client от пользователя
+            # Request updated client data from the user
             name, surname, email = self.view_client.get_client_input()
-            # Вызов метода из класса Model для обновления client
+            # Call a method from the Model class to update the client
             success = (self.model_client.update_client(client_id, name, surname, email))
 
-            # Отображение сообщения о результате операции
+            # Display a message about the result of the operation
             if success:
                 self.view_client.show_client_message("Client updated successfully!")
             else:
@@ -40,18 +40,25 @@ class ControllerClient:
 
     def delete_client(self):
         client_id = self.view_client.get_client_id()
-        if self.model_client.delete_client(client_id):
-            self.view_client.show_client_message("Client deleted successfully!")
+
+        # Check if there is a reservation with the specified ID
+        client_exists = self.model_client.check_client_existence(client_id)
+
+        if client_exists:
+            if self.model_client.delete_client(client_id):
+                self.view_client.show_client_message("Client deleted successfully!")
+            else:
+                self.view_client.show_client_message("Failed to delete client.")
         else:
-            self.view_client.show_client_message("Failed to delete client.")
+            self.view_client.show_client_message("Client with the specified ID does not exist.")
 
     def create_client_sequence(self):
-        # Виклик методу create_client_sequence з класу ModelClient
+        # Call the create_client_sequence method from the ModelClient class
         self.model_client.create_client_sequence()
         self.view_client.show_client_message("Client sequence created successfully!")
 
     def generate_rand_client_data(self, number_of_operations):
-        # Виклик методу generate_rand_client_data з класу ModelClient
+        # Call the generate_rand_client_data method from the ModelClient class
         success = self.model_client.generate_rand_client_data(number_of_operations)
 
         if success:
@@ -60,7 +67,7 @@ class ControllerClient:
             self.view_client.show_client_message("Failed to generate clients.")
 
     def truncate_client_table(self):
-        # Викликаємо метод відповідного model
+        # Call the method of the corresponding model
         success = self.model_client.truncate_client_table()
 
         if success:
